StoreController 의 deleteStore method 및 관련 service logic 수정, 그로 인한 myStore file 수정

+++ The_seat/src/main/java/com/star/seat/store/service/StoreServiceImpl.java
@@ -430,10 +430,9 @@ public class StoreServiceImpl implements StoreService{
 	
 	// 매장 정보를 삭제하는 method
 	@Override
-	public void deleteStore(HttpServletRequest request) {
+	public Map<String, Object> deleteStore(StoreDto dto, HttpServletRequest request) {
 		String email=(String)request.getSession().getAttribute("email");
 
-		StoreDto dto = new StoreDto();
 		SeatDto sDto = new SeatDto();
 		MenuDto mDto = new MenuDto();
 		ReviewDto rDto = new ReviewDto();
@@ -442,14 +441,13 @@ public class StoreServiceImpl implements StoreService{
 		dto.setOwner(email);
 		oDto.setEmail(email);
 
-		// 해당 email로 만든 모든 상점의 list를 가져옴
-		List<StoreDto> list = dao.getMyStores(email);
-		for(StoreDto tmp:list) {
+		Map<String, Object> map = new HashMap<>();
+		if(dto.getNum() != 0) {
 			// 각 상점 번호로 자리, 메뉴, 리뷰, 주문 정보를 다룰 수 있도록 세팅하고
-			sDto.setNum(tmp.getNum());
-			mDto.setStoreNum(tmp.getNum());
-			rDto.setStoreNum(tmp.getNum());
-			oDto.setNum(tmp.getNum());
+			sDto.setNum(dto.getNum());
+			mDto.setStoreNum(dto.getNum());
+			rDto.setStoreNum(dto.getNum());
+			oDto.setNum(dto.getNum());
 			
 			// 매장 정보를 지우고
 			dao.deleteStore(dto);
@@ -461,8 +459,36 @@ public class StoreServiceImpl implements StoreService{
 			rDao.deleteAllReview(rDto);
 			// 매장에서 주문했던 내역도 지움
 			oDao.deleteAllOrder(oDto);
-			// 해당 유저가 다른 곳에서 주문했던 내역도 지움
-			oDao.deleteEmailOrder(oDto);
+			
+			map.put("isDeleted", true);
+		} else {
+			// 해당 email로 만든 모든 상점의 list를 가져옴
+			List<StoreDto> list = dao.getMyStores(email);
+			for(StoreDto tmp:list) {
+				// 각 상점 번호로 자리, 메뉴, 리뷰, 주문 정보를 다룰 수 있도록 세팅하고
+				dto.setNum(tmp.getNum());
+				sDto.setNum(tmp.getNum());
+				mDto.setStoreNum(tmp.getNum());
+				rDto.setStoreNum(tmp.getNum());
+				oDto.setNum(tmp.getNum());
+				
+				// 매장 정보를 지우고
+				dao.deleteStore(dto);
+				// 매장 자리 정보도 지움
+				stDao.seatDelete(sDto);
+				// 매장 메뉴 정도도 지움
+				mDao.deleteAllMenu(mDto);
+				// 매장 리뷰 정보도 지움
+				rDao.deleteAllReview(rDto);
+				// 매장에서 주문했던 내역도 지움
+				oDao.deleteAllOrder(oDto);
+				// 해당 유저가 다른 곳에서 주문했던 내역도 지움
+				oDao.deleteEmailOrder(oDto);
+				
+				map.put("isDeleted", true);
+			}
 		}
+		
+		return map;
 	}
 }
